requirements.txt
flask
gunicorn
flask-login
flask-sqlalchemy
werkzeug
Pillow
pypdf
python-docx
google-generativeai
python-dotenv
pinecone